Moved logic into hooks
